requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.0.0=pyhd8ed1ab_0
aiohttp=3.9.1=py39hd1e30aa_0
aiosignal=1.3.1=pyhd8ed1ab_0
alsa-lib=1.2.10=hd590300_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.3=pyhd8ed1ab_0
attr=2.5.1=h166bdaf_1
attrs=23.1.0=pyh71513ae_1
blas=1.0=openblas
blinker=1.7.0=pyhd8ed1ab_0
brotli=1.0.9=he6710b0_2
brotli-python=1.1.0=py39h3d6467e_1
bzip2=1.0.8=hd590300_5
c-ares=1.24.0=hd590300_0
ca-certificates=2023.12.12=h06a4308_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.2=pyhd8ed1ab_0
cairo=1.18.0=h3faef2a_0
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.16.0=py39h7a31438_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=py39hdb19cb5_0
cryptography=41.0.7=py39he6105cc_1
cuda-version=11.8=h70ddcb2_2
cudatoolkit=11.8.0=h4ba93d1_12
cudnn=8.8.0.121=hcdd5f01_4
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.8.0=py39h3d6467e_1
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=h6a678d5_0
flask=2.2.5=py39h06a4308_0
flatbuffers=23.5.26=h59595ed_1
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.2=h14ed4e7_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=py39hd1e30aa_0
gast=0.5.4=pyhd8ed1ab_0
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
glib=2.78.3=hfc55251_0
glib-tools=2.78.3=hfc55251_0
google-auth=2.25.2=pyhca7485f_0
google-auth-oauthlib=1.0.0=pyhd8ed1ab_1
google-pasta=0.2.0=pyh8c360ce_0
graphite2=1.3.14=h295c915_1
grpcio=1.54.3=py39h227be39_0
gst-plugins-base=1.22.7=h8e1006c_1
gstreamer=1.22.7=h98fc4e7_1
h5py=3.10.0=nompi_py39h2c511df_101
harfbuzz=8.3.0=h3d44ed6_0
hdf5=1.14.3=nompi_h4f84152_100
icu=73.2=h59595ed_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.0.0=pyha770c72_0
importlib_metadata=7.0.0=hd8ed1ab_0
importlib_resources=6.1.0=py39h06a4308_0
ipykernel=6.26.0=pyhf8b6a83_0
ipython=8.18.1=pyh707e725_3
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=py39h06a4308_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.5.0=py39hf3d152e_0
keras=2.14.0=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.21.2=h659d440_0
lame=3.100=h7b6447c_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libabseil=20230125.3=cxx17_h59595ed_0
libaec=1.1.2=h59595ed_1
libblas=3.9.0=20_linux64_openblas
libcap=2.69=h0f662aa_0
libcblas=3.9.0=20_linux64_openblas
libclang=15.0.7=default_hb11cfb5_4
libclang13=15.0.7=default_ha2b6cf4_4
libcups=2.3.3=h4637d8d_4
libcurl=8.5.0=hca28451_0
libdeflate=1.19=hd590300_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=hd590300_2
libevent=2.1.12=hdbd6064_1
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libflac=1.4.3=h59595ed_0
libgcc-ng=13.2.0=h807b86a_3
libgcrypt=1.10.3=hd590300_0
libgfortran-ng=13.2.0=h69a702a_3
libgfortran5=13.2.0=ha4646dd_3
libglib=2.78.3=h783c2da_0
libgomp=13.2.0=h807b86a_3
libgpg-error=1.47=h71f35ed_0
libgrpc=1.54.3=hb20ce57_0
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=20_linux64_openblas
libllvm15=15.0.7=hb3ce162_4
libnghttp2=1.58.0=h47da74e_1
libnsl=2.0.1=hd590300_0
libogg=1.3.5=h27cfd23_1
libopenblas=0.3.25=pthreads_h413a1c8_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h753d276_0
libpq=16.1=h33b98f1_7
libprotobuf=3.21.12=hfc55251_2
libsndfile=1.2.2=hc60ed4a_1
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.44.2=h2797004_0
libssh2=1.11.0=h0841786_0
libstdcxx-ng=13.2.0=h7e041cc_3
libsystemd0=255=h3516f8a_0
libtiff=4.6.0=ha9c0a0a_2
libuuid=2.38.1=h0b41bf4_0
libvorbis=1.3.7=h7b6447c_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.6.0=hd429924_1
libxml2=2.12.3=h232c23b_0
libzlib=1.2.13=hd590300_5
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.1=pyhd8ed1ab_0
markupsafe=2.1.3=py39hd1e30aa_1
matplotlib=3.8.0=py39h06a4308_0
matplotlib-base=3.8.0=py39h1128e8f_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
ml_dtypes=0.2.0=py39hddac248_2
mpg123=1.32.3=h59595ed_0
multidict=6.0.4=py39hd1e30aa_1
munkres=1.1.4=py_0
mysql-common=8.0.33=hf1915f5_6
mysql-libs=8.0.33=hca2cd23_6
nccl=2.19.4.1=h6103f9b_0
ncurses=6.4=h59595ed_2
nest-asyncio=1.5.8=pyhd8ed1ab_0
nspr=4.35=h6a678d5_0
nss=3.96=h1d7d5a4_0
numpy=1.26.2=py39h474f0d3_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.5.0=h488ebb8_3
openssl=3.2.0=hd590300_1
opt_einsum=3.3.0=pyhc1e730c_2
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=py39hddac248_0
parso=0.8.3=pyhd8ed1ab_0
pcre2=10.42=hebb0a14_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.1.0=py39had0adad_0
pip=23.3.2=pyhd8ed1ab_0
pixman=0.42.2=h59595ed_0
platformdirs=4.1.0=pyhd8ed1ab_0
ply=3.11=py39h06a4308_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.21.12=py39h227be39_0
psutil=5.9.7=py39hd1e30aa_0
ptyprocess=0.7.0=pyhd3deb0d_0
pulseaudio-client=16.1=hb77b528_5
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pyhd8ed1ab_0
pyasn1-modules=0.3.0=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_0
pyopenssl=23.3.0=pyhd8ed1ab_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.15.10=py39h6a678d5_0
pyqt5-sip=12.13.0=py39h5eee18b_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.13=h2660328_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-flatbuffers=23.5.26=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.9=4_cp39
pytz=2023.3.post1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyzmq=25.1.2=py39h8c080ef_0
qt-main=5.15.8=h450f30e_18
re2=2023.03.02=h8c504da_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.3.0=py39h1128e8f_0
scipy=1.11.4=py39heeff2f4_0
setuptools=68.2.2=pyhd8ed1ab_0
sip=6.7.12=py39h6a678d5_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h9fff704_0
sqlite=3.44.2=h2c6b66d_0
stack_data=0.6.2=pyhd8ed1ab_0
tensorboard=2.14.1=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py39hd4f0224_1
tensorflow=2.14.0=cuda118py39hc3a5e0e_0
tensorflow-base=2.14.0=cuda118py39hbe86951_0
tensorflow-estimator=2.14.0=cuda118py39h695a165_0
termcolor=2.3.0=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.0.1=py39h06a4308_0
tornado=6.3.3=py39hd1e30aa_1
traitlets=5.14.0=pyhd8ed1ab_0
typing-extensions=4.9.0=hd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=2.1.0=pyhd8ed1ab_0
wcwidth=0.2.12=pyhd8ed1ab_0
werkzeug=3.0.1=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
wrapt=1.14.1=py39hb9d737c_1
xcb-util=0.4.0=hd590300_1
xcb-util-image=0.4.0=h8ee46fc_1
xcb-util-keysyms=0.4.0=h8ee46fc_1
xcb-util-renderutil=0.3.9=hd590300_1
xcb-util-wm=0.4.1=h8ee46fc_1
xkeyboard-config=2.40=hd590300_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.7=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xf86vidmodeproto=2.3.1=h7f98852_1002
xorg-xproto=7.0.31=h27cfd23_1007
xz=5.2.6=h166bdaf_0
yarl=1.9.3=py39hd1e30aa_0
zeromq=4.3.5=h59595ed_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
